<refactor> 역할에 따라 패키지 구조 수정
- session과 관련된 클래스를 webserver 패키지로 이동
